Finish hooking up to existing backend
