+++ react/src/pages/admin/Admin.tsx
@@ -5,7 +5,7 @@ import Tabs from "@material-ui/core/Tabs";
 import Tab from "@material-ui/core/Tab";
 import Typography from "@material-ui/core/Typography";
 import Box from "@material-ui/core/Box";
-import UserList from "./NewUserList";
+import UserList from "./UserList";
 import ManagedUploaders from "./ManagedUploaders";
 
 interface TabPanelProps {
